added migrations files for review, separated movie and reviews in separate files, added reviews to both movie list and detail
