+++ coding_challenge/movies/models/__init__.py
@@ -1,3 +1,4 @@
 from .movie import Movie
+from .review import Review
 
-__all__ = ["Movie"]
+__all__ = ["Movie", "Review"]
