+++ coding_challenge/movies/models/movie.py
@@ -5,8 +5,3 @@ class Movie(models.Model):
     title = models.CharField(max_length=100, unique=True)
     runtime = models.PositiveIntegerField()
     release_date = models.DateField()
-
-class Review(models.Model):
-    movie = models.ForeignKey(Movie, on_delete=models.CASCADE, related_name='reviews')
-    reviewer_name = models.CharField(max_length=100)
-    rating = models.IntegerField(choices=[(i, i) for i in range(1, 6)])
