+++ coding_challenge/movies/models/review.py
@@ -1,0 +1,7 @@
+from django.db import models
+from movies.models import Movie
+
+class Review(models.Model):
+    movie = models.ForeignKey(Movie, on_delete=models.CASCADE, related_name='reviews')
+    reviewer_name = models.CharField(max_length=100)
+    rating = models.IntegerField(choices=[(i, i) for i in range(1, 6)])
