+++ coding_challenge/movies/serializers/__init__.py
@@ -1,3 +1,3 @@
 from .movie_serializer import MovieSerializer
 
-__all__ = ["MovieSerializer"]
+__all__ = ["MovieSerializer", "ReviewSerializer"]
