+++ coding_challenge/movies/serializers/movie_serializer.py
@@ -1,3 +1,4 @@
+from coding_challenge.movies.serializers.review_serializer import ReviewSerializer
 from rest_framework import serializers
 
 from movies.models import Movie
@@ -9,7 +10,10 @@ class RuntimeFormattedField(serializers.Field):
         return f"{hours}:{minutes:02d}"
 
 class MovieSerializer(serializers.ModelSerializer):
+
     runtime_formatted = RuntimeFormattedField(source='runtime')
+    reviews = ReviewSerializer(many=True, read_only=True)
+    
     class Meta:
         model = Movie
         fields = (
@@ -18,4 +22,5 @@ class MovieSerializer(serializers.ModelSerializer):
             "runtime",
             "runtime_formatted",
             "release_date",
+            "reviews"
         )
