+++ coding_challenge/movies/serializers/review_serializer.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+
+from movies.models import Review 
+
+class ReviewSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Review
+        fields = ('reviewer_name', 'rating')
