+++ coding_challenge/movies/urls.py
@@ -4,5 +4,5 @@ from movies.views import MovieListView, MovieDetail
 
 urlpatterns = [
     path("", MovieListView.as_view(), name="MovieListView"),
-    path('<int:pk>/', MovieDetail.as_view(), name='movie-detail'),
+    path('<int:pk>/', MovieDetail.as_view(), name='MovieDetail'),
 ]
