+++ coding_challenge/movies/views/__init__.py
@@ -1,4 +1,4 @@
 from .movie_list import MovieListView 
-from .movie_list import MovieDetail 
+from .movie_detail import MovieDetail 
 
 __all__ = ["MovieListView", "MovieDetail"]
