+++ coding_challenge/movies/views/movie_list.py
@@ -1,13 +1,8 @@
-from rest_framework.generics import ListCreateAPIView, RetrieveUpdateDestroyAPIView
+from rest_framework.generics import ListCreateAPIView
 
 from movies.models import Movie
 from movies.serializers import MovieSerializer
 
-
 class MovieListView(ListCreateAPIView):
     queryset = Movie.objects.order_by("id")
-    serializer_class = MovieSerializer
-
-class MovieDetail(RetrieveUpdateDestroyAPIView):
-    queryset = Movie.objects.all()
     serializer_class = MovieSerializer
